+++ src/Controllers/Weather/Get.php
@@ -10,7 +10,6 @@ use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use QL\CJarvis\MVC\libs\ControllerInterface;
 use QL\CJarvis\MVC\libs\Template;
-use QL\CJarvis\MVC\models\Weather;
 
 class Get implements ControllerInterface
 {
@@ -25,18 +24,7 @@ class Get implements ControllerInterface
         ResponseInterface $response,
         Template $template
     ) {
-        $parseBody = $request->getParsedBody();
-
-        $model = new Weather();
-
-        $weather = $model->weatherMan();
-
-        //mail('chris.jarvs@gmail.com', 'test', 'test bruh');
-
-        return $template->render([
-            'CITY' => $weather['city'],
-            'STATE' => $weather['state'],
-            'DEGREE_F' => $weather['temp_f'],
-            ]);
+        
+        return $template->render([]);
     }
 }